requirements.txt
SpeechRecognition
pyaudio
gTTS
google-generativeai
python-dotenv
streamlit

-e .